pyproject.toml
[project]
name = "dynamic-fixed"
version = "0.1.0"
description = "Job role analysis engine with FastAPI UI and configurable LLM pipelines"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.110.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
    "pydantic>=2.7.0",
    "pytest>=8.4.2",
    "pyyaml>=6.0",
    "sentence-transformers>=3.0.0",
    "uvicorn>=0.37.0",
]

[project.optional-dependencies]
embeddings = [
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.0",
]
dev = [
    "pytest>=8.3.0",
]
